Add native custom derive on nightly Rust
Stable Rust will still use transpilation via build script
but on nightly native custom derive will be used so we can
benefit from proper error locations during development.
