+++ crates/easter/Cargo.toml
@@ -9,10 +9,14 @@ homepage = "https://esprit.surge.sh"
 repository = "https://github.com/dherman/esprit"
 build = "build.rs"
 
+[features]
+default = ["derive-context"]
+nightly = ["derive"]
+
 [dependencies]
 joker = { version = "0.0.5", path = "../joker" }
 tristate = "0.1.1"
+derive = { version = "0.0.1", path = "../derive", optional = true }
 
 [build-dependencies]
-syn = { version = "0.10.5", features = ["parsing", "printing", "expand"] }
-quote = "0.3.10"
+derive-context = { version = "0.0.1", path = "../derive-context", optional = true }
