+++ crates/easter/src/lib.rs
@@ -1,6 +1,18 @@
+#![cfg_attr(feature = "nightly", feature(proc_macro))]
+
 extern crate joker;
 extern crate tristate;
 
+#[cfg(feature = "nightly")]
+#[macro_use]
+extern crate derive;
+
+#[cfg(feature = "nightly")]
+macro_rules! pub_mod {
+    ($name:ident) => (pub mod $name;)
+}
+
+#[cfg(not(feature = "nightly"))]
 macro_rules! pub_mod {
     ($name:ident) => (pub mod $name {
         include!(concat!(env!("OUT_DIR"), "/", stringify!($name), ".rs"));
